seperate suervey response modal component
