+++ app/assets/javascripts/angular/components/organizationOverviewSurveyResponses/surveyResponsesModal.component.js
@@ -1,0 +1,16 @@
+import modalTemplate from './surveyResponsesModal.html';
+
+angular.module('missionhubApp').component('surveyResponseModal', {
+    template: modalTemplate,
+    bindings: {
+        dismiss: '&',
+    },
+    controller: function($scope, localStorageService) {
+        var vm = this;
+
+        $scope.closeModal = function() {
+            localStorageService.set('newSurveyResponseModal', true);
+            vm.dismiss({ $value: 'cancel' });
+        };
+    },
+});
